first

+++ cfast_slic.pxd
@@ -7,10 +7,10 @@ cdef extern from "fast-slic-common.h":
     ctypedef struct Cluster:
         uint16_t y
         uint16_t x;
-        uint8_t r;
-        uint8_t g;
-        uint8_t b;
-        uint8_t reserved[1]
+        uint16_t r;
+        uint16_t g;
+        uint16_t b;
+        uint16_t reserved[1]
         uint16_t number
         uint8_t is_active
         uint32_t num_members
